pyproject.toml
[build-system]
build-backend = 'setuptools.build_meta'
requires = ['setuptools']

[project]
name = 'server'
version = '0.0.1'
dependencies = [
    'aiofiles',
    'cryptography',
    'fastapi',
    'loguru',
    'ormar[sqlite]',
    'pillow',
    'pydantic',
    'python-multipart',
    'uvicorn[standard]',
]

[[project.authors]]
name = 'Dragon-GCS'
email = '878173121@qq.com'

[project.optional-dependencies]
dev = [
    'httpx',
    'requests',
    'ruff',
]
[tool.setuptools]
packages = ['server']

[tool.ruff]
line-length = 100
